fix(translator) Allow curried functions to be translated to functions (#1739)
Allow curried functions to be translated into functions, which lets us convert
them to identity functions in impossible queries. This isn't strictly necessary,
as the curried functions in clickhouse accept NULL and return NULL. Less is more
though when it comes to queries.

There was some tech debt preventing this, but it seems to be resolved now. Maybe
the great Typing of 2021 fixed it.