fix(errors): Fix group deletion (#1744)
Fixes a bug causing events to not be properly deleted in the errors table
due to an incorrect primary hash value.

The deletion being inserted had the wrong sorting key, causing the row
not to be properly deduplicated. This seems to have gone unnoticed for some
time since the error is only noticable once excluded groups for the project
are expired from Redis cache and those groups are not being explicitly excluded
in queries by the PostReplacementConsistencyEnforcer.

This bug only applies to the errors table, as `primary_hash` was not part of
the sorting key on the events (sentry_local) table.