feat(errors): Add type optimization for errors storage (#1748)
The expression `type != transaction` is frequently included in queries
in order to filter rows because the events table includes transaction events.
Since the errors table never includes these events to begin with, this condition
will always be met. This optimizer rewrites `type != transaction` to `1` if it is
present when running the errors storage, so that the ClickHouse query does not need
to unnecessarily access the `type` column.