feat: Improve UUID processing so it can be extended to other columns (#1754)
This processor combines the functionality of what is currently split across the
EventIdColumnProcessor and the UUIDColumnProcessor. It gives us the ability to
easily add support and optimizations for UUID columns without having a specific
processor for each column. It also makes it impossible to accidentally apply one
of the processors but not the other one, or apply them in the wrong order which
will result in broken behavior. Now you can no longer apply a UUID processor
without the optimization for conditions as well.

The new processor is only applied to the errors storage to begin with.