fix(reprocessing): Tombstone for particular ReplacerState (#1734)
When an event is inserted over in Snuba through reprocessing, we use
redis to make sure the old row is excluded from queries until table
merge is done. However, with the new errors table we have a problem: An
event inserted over the old one may not have the same PK/sortkey as the
old event, because primary_hash is part of PK. That means table merge
will see those two events as not-to-be-merged even though they have the
same event ID.

We actually don't really know what that does in the product, having two
events queryable with the same event ID. It seems like this case can
also be hit from the outside by simply sending two events that have the
same event ID, but group differently. So arguably this needs to be fixed
for cases beyond reprocessing, but reprocessing is by far the most
common case where this happens, so we just fix it there.

This change adds a new flag to the tombstone-events message to tell
Snuba to only run this replacer on the new errors table. Sentry will
then emit this message before it reinserts an event as part of
reprocessing, and has detected that the primary hashes are different.