feat(cdc) Expand BatchWriter to support bulk load (#1764)
BatchWriter and HTTPWriteBatch are designed to store individual JSON rows. Even the old snapshot loader implementation was sending rows as JSON rows (with bigger chunks).

This would not work well to ingest into Clickhouse a massive pre-formatted gzip file.
This PR adds some extensions to the BatchWriter to be effective at ingesting big chunks of binary data as opposed to individual rows. Specifically:

Introduce InsertStatement to capture the parameters needed. Before this, individual parameters were passed through. Now we need to pass encoding and format parameters, so this makes signatures cleaner
Allows the buffer_size to be configured in HTTPWriteBatch. By default the buffer is a SimpleQueue which is unbounded. This is not great when we try to upload a 7GB file. So a bounded buffeer is represented with a Queue with a predefined size.
It also expose the bulk_writer on the cluster