feat(cdc) Support pre processed snapshot in the bulk_load script (#1765)
The previous PR support big binary file in the BatchWriter. This one adapt the bulk_load script to deal with that change.

Specifically:

separates the functionality of the script in parse and load and load preprocessed data
split get_table intoget_parsed_table_file and get_preprocessed_table_file
do the same for the load.