feat: Add a type condition enforcer
The type condition enforcer ensures that transactions are never
returned from the events entity. This is an issue when the events
entity is picked together with the legacy events storage, and Sentry works
around this by including the `type != transaction` condition on
most queries to events.

This change will allow us to stop sending that condition on all queries from
Sentry, as Snuba will automatically apply if it is required. It also allows us
to remove the TypeConditionOptimizer that operates on the errors storage in order
to optimize that unncessary condition away.