Add concurrent db queries (#1775)
We were used to have a gauge metric that represented the global count of concurrent queries.
This was dropped, not sure why, long ago. But it would still be useful especially to understand the trend, an upward trend in concurrent queries can point to a problem about to happen (like we are about to hit the global rate limit or to a query execution performance issue).

This resurrect a limited version of that metric (counts the queries right before sending them to Clickhouse).
A future version will also keep track of the queries before the processing phase, which means before caching and rate limiting.