perf(ast) Avoid copying identical expression when transforming (#1778)
Running the transform method over large nested expression is something done many times per query as it is the basic building block of any expression transformation.
For example it is done during translation and for every query processor.
Most simple expression do not get modified though, right now, nested expressions (functions, curried functions, subscriptables and lambdas) instantiate a new version of themselves even when the children do not change. This is a fairly important amount of object allocations we do not need as expressions are immutable.