Remove useless deepcopy (#1779)
In the past we had to copy the content of the request before doing any query processing so that we could log the query body provided by the user in the querylog.

This was before the AST and before the separation into logical and physical query. The ast was actually the same Mapping used as query body so it was modified during the processing.
This is not the case anymore as the query body is never touched after parsing, the settings object and the query object are not used for the sake of the query logging (we extract information only from the physical queries after processing), so we can drop this copy.

The query object is much bigger than the query body itself. When profiling Snuba cpu usage it turned out that this deepcopy should have a visible impact.