perf(replacer) Make the event tombstone query condition more performant (#1784)
This makes two changes to the process_tombstone_events method in the replacer:

Avoid PREWHERE replaceAll(toString(event_id), '-', '') IN and replace it with cityHash64(toString(event_id)) IN (cityHash64(...)) when finding events
allow the client to pass a time range to the replacer, so we can restrict the query. The time range is the one to execute the query onto, which means it has to be computed by the client when passing a list of events.

This is supposed to make replacements generally faster and to make it cheaper to run batch deletion on a large number of events.