feat: Mapping column promoter does not cast to string by default
The mapping column promoter now only promotes a column and does not
attempt to fix it's type by simply casting to string. We need to keep
the old behavior around temporarily as it's used in the legacy events
table - once that is gone the option to cast the type will no longer
be required here.

The previous behavior caused issues with the UUIDColumnProcessor as the
expected query no longer matched what the processor expected if a column
was a promoted one due to the insertion of the extra toString() function.