feat: Add HexIntColumn processor and promote the span_id column (#1796)
Now we are no longer using the span_id from contexts[trace.span_id] but
getting it directly from span_id in all cases. As a follow up the UUID processor
should also be converted to use the BaseTypeConverter. We could also completely
remove the span_id from the contexts array if we wanted to.