feat: Errors migration (#1801)
Convert the errors backfill script into a migration. Also adds a migration
to truncate the events table once the data is copied to the new table.

This also flips the switch on the errors rollout so that we are now using the
new table. This will affect onpremise, Sentry dev and CI environments.