ref: Remove ability to pick storages based on query referrer
Now that the new storage is rolled out in production, we can remove the
code that allows switching between storages for specific queries as well
as the ability to run certain queries on multiple storages and compare the results.

We are still keeping the option to switch between the events and errors/discover
storages, but it's either one or the other now; we no longer need the more fine
grained controls we used for rollout.