feat(snql) Add array join support
The ability to do basic array join queries was left out of SnQL on an assumption
that the product could use the "arrayJoin" function instead. However that's not
practical because there needs to be a way to do an array join without adding an
expression to the SELECT or GROUP BY clauses. Note that this does NOT solve the
issue of the legacy API implicitly expanding conditions on array columns.