feat: The "events" topic should be created with LogAppendTime
This change creates the topics for the errors and transactions storages
with LogAppendTime configuration. This is required for subscriptions, as
subscriptions on events and transactions require time to move monotonically
on the topics they subscribe to. By default, the storages share the same "events"
topic, but this configuration can be changed via settings.

This solution feels like a pretty big hack, but I struggled to come up with
a better alternative since there is currently nothing to enforce that these
storages even share a topic. Other suggestions are welcome.