feat: The "events" topic should be created with LogAppendTime (#1814)
This change creates the topics for the errors and transactions storages
with LogAppendTime configuration. This is required for subscriptions, as
subscriptions on events and transactions require time to move monotonically
on the topics they subscribe to. By default, the storages share the same "events"
topic, but this configuration can be changed via settings.