fix: Prevent all conditions from being dropped from a query (#1822)
This is a quick fix to prevent the entire condition expression
from being mapped to null by the DefaultIfNullFunctionMapper. It
ensures that mapping of and/or/not/xor expressions is skipped
even if a match for the identity(...) expression is found.

We sbould probably follow up with a more optimal solution that does
not replace any expression that contains identity(...) with
identity(NULL).