ref(snql) Refactor required column/time condition code
When support for SnQL subscriptions is added, it's going to be necessary to
easily skip time condition processing on the SnQL query. Refactor how the
conditions are checked to remove the checks from the entity class (they never
should have been there to begin with). Separate out the time condition
processing so that in a subsequent PR it can be simply skipped. Move the other
logic into places that make more sense going forward.