fix(splitter) Skip the splitter if the order by is a function (#1833)
It's possible for the minimal query to break if the orderby is a function on a
column instead of a bare column. Clickhouse throws an exception along the lines
of `Not found column f(x) in block` if f(x) is not in the select clause. Since
the minimal query replaces the select clause, this failure happens. Skip the
splitter entirely in this case.