ref: Simplify Kafka topic configuration (#1820)
Previously it was possible to specify the default, commit-log and replacement
topic name and broker configuration for any storage via settings independently.
There was nothing to enforce that certain storages should share a topic, even
though Snuba might depend on this. If you wanted to override the default "events"
topic for example, you needed to specify the new name 4 times under each associated
storage - events, errors, transactions, spans.

This patch deprecates the QUERIES_TOPIC, STORAGE_TOPICS and STORAGE_BROKER_CONFIG
settings in favor of KAFKA_TOPIC_MAP and KAFKA_BROKER_CONFIG, which allows the user to
map a default topic name to a new one and to custom configuration.
The number of topics and their relationship to storages is no longer configurable.

Note: STORAGE_TOPICS and STORAGE_BROKER_CONFIG are still temporarily supported, and
will be removed in a follow up.