ref: Simplify storage selection
Since we no longer need to run queries on both storages
at once, we can simplify storage selection on the events
and discover entities.

This also prevents failed healthchecks on unused storages.
Only the errors storage is checked if ERRORS_ROLLOUT_ALL is
set to true, otherwise only the events storage is checked.