fix(query) Add a failsafe processor to avoid running clickhouse queries that are missing conditions (#1823)
Today we found an issue where a bug in a translator (the one or impossible queries) was wiping the whole condition clause in the query replacing it with identity(NULL) this query never returns anything so there is no concerning impact but that bug is scary as it removed essential conditions from the query (project id and timestamp).

This adds a failsafe solution in to prevent snuba from running queries that lost the project condition for any reason. The check should always pass, this is just meant to reduce the blast radius of a potential query processing bug that would compromise the query structure.

Ideally we should make changes to the query structure so that removing project_id conditions by mistake would be impossible but that is tricky and will take longer to build as there are use cases where the project_id condition is supposed to actually be removed from the query (when it is moved to the prewhere).