feat(metrics) Storage and migration for the metrics prototype (#1812)
Adds a storage for set bucketed metrics, for the metrics prototype.
This storage is disabled by default, which means it will not be created for self hosted users and it needs to be activated for the migrations system to create it.

This is meant to allow the ingest team to build the prototype of the metrics based sessions product.
This part only covers set metrics (distributions and counters will be in a follow up otherwise this PR would become gigantic).
Metrics are written in bucketed form in a single table metrics_buckets_local then there would be multiple materialized views and aggregated tables, one per metric type.
Each aggregated table can contain multiple granularities. At this time there are only minute based aggregations.

Some additional details in comments in line.