feat: Subscription result Kafka topic can be configured via settings (#1859)
This is an attempt to align the way subscription result topics are configured
with the rest of the Kafka topics.

Previously any storage could be subscribed to (though it may not work properly).
Now a storage defines whether it can be subscribed to or not via
`stream_loader.subscription_result_topic`. It is no longer necessary to pass the
--result-topic into the subscriptions consumer, the topic name will come from
the default or settings configuration like all other Kafka topics.

The result topics will also now get created when `snuba bootstrap` is run.