feat: Introduce metrics class for streaming consumer
The purpose of this change is to remove the dependency that currently
exists between streaming consumer code (specifically the stream processor
and parallel processing steps) to Snuba's metrics module.

This is solved by introducing a separate metrics abstract class (this
will eventually move out of Snuba's codebase into the streaming library)
as well as an adapter (this will stay in Snuba) to convert Snuba's metrics
backend to the metrics class expected by the future streaming library. We
also copy over the Gauge implementation as this can no longer depend on the
version in snuba.utils.metrics.

The `metrics` argument no longer needs to be passed to StreamProcessor or ParallelTransformStep.
If this is not present a default (which records nothing) will be used.