feat(dev) Introduce settings to enable/disable dev data model elements (#1862)
When introducing a new dataset (or a new entity/storage in an existing dataset), we generally want to have the code in a functioning state before the schemas are applied and existing in all environments. We need to do development on the draft migrations before we have them applied to prod.
We do not have a good way to isolate parts of the data model that are still under development from the application in prod if we need to commit the code.
(The DISABLED_DATASET field helps but it does not have the right granularity and it does not disable a storage set).

This introduces a flag in the settings that turns on/off features under development. Features under development can be datasets, entities and storages. These are declared as "dev" features in code in their factory and they would be invisible if the flag is off.
Of course this flag will be off in all environments except when a developer is relying on these features.

This change would not have been enough by itself to isolate a dataset (and solve the deployment issue we had with metrics). Storages are initialized when the module is loaded, thus in the current implementation they would try to instantiate a cluster for their storage set during initialization.

This is wrong. StorageSetKey is part of data model configuration, while the cluster config is specific to the environment, the data model config should always be the same everywhere, while the cluster config depends on the environment. So this also decouples the cluster initialization from the data model initialization. The data model initialization only stores the storage key now.

This would have still not been enough. To fully hide a new dataset we would also need to disable storage sets which are not assigned to a cluster yet. Thus the change to the clusters.py file.