feat(snql) Boilerplate for SnQL subscriptions (#1866)
We will have to support SnQL in subscriptions (even if we do not run into alerts on joins, this would be a requirement to deprecate the legacy language).

This adds some boilerplate by making SubscriptionData abstract (legacy and SnQL will extend it) and adapts all the tests to instantiate LegacySubscriptionData instead SubscriptionData.

The next step will introduce the SnQL subclass.
The idea is to have:

two formats for storing the subscription query (this is mandatory as the language is different). I still need to decide whether we will have two SubscriptionStore instances with two different codecs, or whether one codec will deal with both. This is my goal in the next PR.
the build_request method will be implement differently between the old and new language as the parsing is different.
Once the Request is built there will be mostly no difference between the two as only the AST is referenced from that point on.