feat: Insert batch encoding is done by message processors
JSON encoding of insert batches should not be done by the streaming
consumer strategy factory, as this code should eventually become generic
and agnostic of encoding formats so that it can be used outside of Snuba.

Now it's the job of the message processor to do the encoding and convert
an InsertBatch into a JSONRowInsertBatch.