feat(snql) Split request validation from parsing. (to allow for SnQL subscriptions) (#1885)
Subscriptions that run legacy language queries are easy to parse. The query body is a dictionary, so the subscription data code can add the timestamp conditions before parsing (which would validate the query structure and reject a query without timestamp conditions).
SnQL is harder as the query body is a string. We won't try to tweak the string to add timestamp conditions, we are going to add them to the AST (where we have tools to do that safely).
This means that we need to do the parsing in two steps. AST generation and validation. Splitting the parsing is easy (Evan already broke the two phases apart).
Now we need to ensure we use the right parser in the subscription code.
Unfortunately query parsing is buried in the request builder together with the schema validation. This does not sound right (validating the schema also parses the query and generates the AST).

So this PR prepares the ground for the process above moving the request generation logic from schema validation (snuba/request/schema) to the build_request code, which now can take a function to parse the query (so we can have three of them: legacy, snql_api, snql_subscriptions). Since we need to pass three different parsers (while the languages are two) we cannot rely anymore on the language to pick the parser).