feat(snql) Parse and run SnQL subscriptions (#1886)
This puts together #1885 and #1866. It adds the build_request method to the SnQLSubscriptionData class. So that a SnQL subscription can be executed.

This parses the query, adds the timestamp conditions, the project condition and the offset condition like for the legacy query. The difference is that these are added on the AST instead of being added to the query body.

Tests are updated to reflect the two types of subscription data classes