feat: Improve metrics API
This is an attempt to improve the way that the metric backend is passed to the
streaming consumer API. The main goals are:

- Avoid the need to pass `metrics` everywhere. This is especially annoying since
Snuba uses both higher and lower level streaming consumer APIs in various places
and the list of places where metrics would need to be passed might be long.

- Make APIs more consistent. Avoid the need for the user of this library to
know about which classes and methods happen to record metrics and hence need
`metrics` passed. This also makes it possible to change the metrics that we
capture in the future without also requiring changes to the API.