feat: Stream processor always attempts to recover from a transport error
Stream processor no longer accepts recoverable_errors as an argument.
Recoverable is now a property of an error and this cannot be configured
by the end user.

The motivation for this change is to simplify the stream processor API,
and not require the user to remember to pass `recoverable_errors=[TransportError]`
to get sensible default behavior.