docs: Tear the old README apart (#1944)
This adds:

an intro page to the sphinx docs.
Add a basic getstarted page with the content that came from README.md
Add a basic dev environment page with the content from README.md
Replace the readme content with an intro and a link to the docs
Move all the legacy query docs in a README in /query/parser
Get started and Environment are just basic at this point. There is a lot more content to add. For this PR I just copied the content from the old README to have the proper structure in place.