feat: Improve consumer metrics API (#1936)
This is an attempt to improve the way that the metrics backend is passed to the
streaming consumer API. The main goals are:

- Avoid the need to pass `metrics` everywhere. This is especially annoying since
Snuba uses both higher and lower level streaming consumer APIs in various places
and the list of places where metrics would need to be passed is long.

- Make APIs more consistent. Avoid the need for the user of this library to
know about which classes and methods happen to record metrics and hence need
`metrics` passed. This also makes it possible to change the metrics that we
capture in the future without also requiring changes to the API.

One potential limitation here is that we can no longer pass different metrics backends
(or paths) to each individual component or function of the streaming API like we used to.
I don't think this is really an issue in Snuba since metrics configuration is done by each CLI
command independently.

Also removes the MetricsWrapper from the consumer library and preserves the original
metric name passed to parallel transform. As a result the metric `snuba.consumer.process.batches_in_progress`
will become `snuba.consumer.batches_in_progress`.