fix(outcomes): Fix outcomes dist migration (#1955)
This fixes a bug where the outcomes dist migrations were incorrectly attempting
to create the `category` column in the outcomes_hourly_local table instead of
outcomes_hourly_dist. As a result a column was missing from the dist table.
There is also no order by on a distributed table so we cannot attempt to run a
MODIFY ORDER BY statement or this migration will just fail.