fix(matchers): Allow Literal matches on None
The current Literal matcher would return a match if the pattern value
was set to None. But there can be cases where we want to match a
pattern value to be None. Added an optional parameter to Literal
matcher called check_none_value which does an explicit check to verify
that the node's value is None. Start using this flag in
DefaultIfNullFuntionMapper.