perf(subscriptions) Test possible message processing improvement by sampling metrics (#1974)
We are logging subscriptions.receive_latency and subscriptions.process_message at every tick of the TickConsumer (which is per message).
This happens in the main consuming thread.

This is a test to measure the impact of these metrics on the consumption time.
In the past we saw there was a visible impact in throughput for a consumer due to these metrics logged for each message.

Probably the gain won't be similar to the one we had on the cdc consumer as the logic is not fully serial in the consumer. Query processing are in separate threads which can use the core while we send the UDP metric to datadog. This is meant to measure that.