feat: Run subscription worker with execute_robust
This change runs subscription worker with the robust ClickHouse execution
strategy. This means that (like replacers) connection issues are retried 3
times and too many concurrent queries will be retried forever with a 1
second delay.

The main motivation for this change is to prevent subscription workers
unnecessarily crashing and restarting when ClickHouse is overloaded.

An alternative strategy would be to have the subscription worker define
it's own retry logic rather than use the default execute_robust behavior.
This may be preferable if we want to be able to define and tweak subscription
specific retry behavior. However this would require the original ClickHouse
error code to be stored on the QueryException so that the subscription worker
can detect which type of error occurred. Since this information isn't present
there currently I picked the seemingly simpler strategy here instead.