feat: Run subscription worker with execute_robust (#1980)
This change runs subscription worker with the robust ClickHouse execution
strategy. This means that (like replacers) connection issues are retried 3
times and too many concurrent queries will be retried forever with a 1
second delay.

The main motivation for this change is to prevent subscription workers
unnecessarily crashing and restarting when ClickHouse is overloaded.