fix(consistent) Have a killswitch we can use in case the amount of consistent queries is too high for clickhouse (#1982)
This is the quickest safeguard in case we still go beyond the max concurrent queries in clickhouse on the nodes that take consistent queries.

This allows compromises consistency (for the percentage of queries configured) by sending to any node.
Hopefully we will not have to use it. But it is better to have it just in case.

The consistent flag means that the query will go to to the same node we write onto so avoiding the replication lag (and use one thread only). Replication lag tends to by tiny: below 10ms), so in case of incident it is probably better being capable to compromise that than to fail large amount of queries.