feat: Initialize Sentry SDK in spawned subprocesses (#1984)
This change initializes the Sentry SDK in the subprocesses
spawned by Snuba consumers. Non-fatal errors and warnings
emitted from these subprocesses should now be sent to our
Sentry instance.

It's possible that Sentry will become a bit more noisy as
a result - we may want to follow up later with tweaks to
optimize what we are actually sending Sentry.