fix(query settings) Create the right instance of query settings (#1985)
We have been creating HTTPRequestSettings instances for subscriptions instead of SubscriptionRequestSettings for quite some time. This is a bug.
IT had no consequence until I added a bypass to reduce the concurrent queries, as the parameters used in the HTTPRequestSettings were the same. I noticed the issue since the bypass generates wrong results removing consistent too often.