perf: Support comparing subscription queries with and without consistent (#1988)
This change adds a way to run and compare the results of subscription queries
without the "consistent" setting to those that currently run in production
with "consistent" set to true. If results turn out not to be significantly
different, it may be worthwhile to turn off consistent for subscriptions
allowing their queries to be spread over all replicas instead of just the
first one.

This can be set for events and transactions subscriptions via the
"event_subscription_non_consistent_sample_rate", and
"transaction_subscription_non_consistent_sample_rate" configuration keys.
Both of these configurations should be set to a value between 0 and 1 where
0 means don't snuplicate any queries (the default) and 1 means run 100% of
queries with both settings and record metrics.

The actual result being returned is unchanged, the secondary query is
only used for recording metrics.