feat(cdc): Apply FINAL clause in CDC tables
Since data in CDC tables are mutable and these tables are fairly smaller
compared to the other tables, we can apply the FINAL clause for queries
to these tables.

Also removed the final member of the Table class from equality tests
since logically any 2 tables with or without the final clause should
be the same.