fix(migrations): Add sharding key to migrations distributed table
We should specify a sharding key for every distributed table whenever possible
(i.e. unless there is a reason why the table absolutely cannot be sharded)
otherwise an exception will be raised when a user attempts to place that storage
set on a ClickHouse cluster with more than one node. Whilst the migrations
table is unlikely to benefit from sharding, there is no reason why a user
can't place it on a sharded cluster if they wish.