fix(sessions) Add a sessions storage without mandatory conditions
We created a sessions entity that could be used to query aggregate numbers
across all organizations, however we didn't create a custom storage as well. The
mandatory condition checks on the old storage were failing, so I created a new
storage that is missing those checks just to be used with that one entity.