fix(snuplicator): Handle exceptions in the primary query (#2013)
Currently the callback function receives a list with all of the query
results. If one of the queries has failed, it is ambiguous whether it is
the primary or some other query that has failed. This change passes
the primary result back separately to the rest of the results. It will
be None if the function has errored.