build(ci): Push a CI-only snuba image to make use of docker caching (#2017)
Docker builds do not seem to be using cache.  Instead, we push (and pull) a CI-only image to GitHub's Container Registry that will be used to make use of docker layer caching.

Changed the test jobs to depend on a new job that builds the snuba CI image. We tag this image with `latest`, `<commit sha>`, and `<branch>`. This should mean that the `Tests and code coverage` jobs and Sentry integration test will have all layers cached when they build the snuba image.