feat(replacements) Run replacement on individual write nodes in cluster environments (#2025)
Replacement queries generally are in this form:
INSERT INTO table SELECT new_fields FROM table WHERE condition
this means that if we run in a cluster environment with multiple shards, we add new shards and clickhouse does not send the entire INSERT SELECT query to write nodes, we could write tombstones on the wrong nodes.

This happens if the query node (where the distributed table is) first executes the select, then receives the results and then execute an INSERT in the distributed table. IT seems to happen at times.
The shards selected for the rows during insertion may not be the same the original rows came from as the sharding logic changed.

This PR mitigates that effect by sending the entire replacement query to each shard individually instead of sending it to the query node on the distributed table.