feat: Move CDC to a separate storage set, cross storage set joins
This change adds a new CDC storage set, which the groupedmessages
and groupassignee storages belong to.

Previously these storages were on the "events" storage set, same as
the errors table. This was wrong because it is impossible for
groupedmessages and groupassignee to share a ClickHouse cluster
and layout (by definition storages on the same storage set share
ClickHouse clusters). Specifically the issue is that errors
being a large table should be able to have more than 1 shard,
but because of the way CDC works, groupedmessages and groupassignees
will not work with more than 1 shard.

In order to solve this problem, we introduce STORAGE_SET_GROUPS,
predefined groups of storage sets that are allowed to be joined on.
Storage sets in a storage set group can have different local nodes but must
share query and distributed nodes. This is checked on initialization
and Snuba will not start if this constraint is not met.

Joins are only allowed between plans with the same storage sets or storage
sets in the same group.