feat(replacement) Load balance INSERT query to storage nodes to avoid overloading the first node of the replica set (#2036)
#2025 sends replacement INSERT queries to individual storage nodes to keep the data consistent after resharding. Before that INSERT queries were going to the query node and the distributed table.

That always preferred the first node of each replica set. As a consequence we are increasing the INSERT statement per second node too much when we have a spike of small replacements in the queue.

This adds a round robin load balancer to pick a different node for each INSERT to spread queries on the cluster.
This load balancer could actually be made generic and added to the native clickhouse module.