feat: Custom zookeeper path for replicated merge tree migrations
CLICKHOUSE_ZOOKEEPER_OVERRIDE allows the Zookeper path for a replicated
merge tree to be customized. Normally this is generated by the migration
system and should not be changed, however under certain circumstances
this is required. Specifically, we have some tables that were created
with different paths prior to the existence of the migration system.
Changing the zookeeper path after a table exists is not easy, so we should
provide a way to override this value as adding a shard to these existing
tables via the migration system will require these paths to match.