test: Fix truncating table in test endpoints and between tests (#2072)
Previously we connected to the query node and truncated the local tables
between each test run. This was basically a no-op on a ClickHouse cluster
since the local tables are not located on the query node and truncating
the distributed table doesn't do anything. Now we connect to each node
of the cluster individually and truncate local tables individually.