Promote span id for discover and errors (#2074) (#2077)
# Problem
We want to reduce the size of the contexts column by removing trace_id and span_id from them but we still want to serve the queries that request `contexts[trace.span_id]`. 

# Solution
Promote `contexts[trace.span_id]` to use `span_id` column instead. Also span_id is queried by a hex string, not the uint64 so we add a HexIntColumnProcessor to each storage to do the translation for us