feat: Commit log topics use LogAppendTime timestamp type
This changes the topic configuration for snuba-commit-log from
CreateTime to LogAppendTime. The purpose of this change is to allow
the timestamp of the commit log topic to be used as the basis for
subscription scheduling (instead of the timestamp from the original
events topic). This change ensures that the message timestamps in the
commit log move monotonically within each partition.

The configuration change in thie PR only applies to newly bootstrapped
topics (i.e. new Sentry installations). We also need to ensure that
the topic configuration gets altered for all existing installations.