feat: Produce message timestamp to commit log headers
This change writes the timestamp of the original Kafka message
in the header of it's commit log topic with the key `message_ts`.
This timestamp is used as the basis for subscription scheduling and
having this in the commit log will enable subscriptions to no longer
need to consume the original topic in future, only the commit log.

This depends on https://github.com/getsentry/arroyo/pull/28 and a
new release of arroyo.